modified error

+++ tools/iso2zip_location.sh
@@ -4,6 +4,15 @@ IMG=$1
 suffix=".iso"
 pkg="zip p7zip-full"
 
+# if there is no parameter, or the parameter is not a file
+if [ -z "$IMG" ]; then
+    read -p "Please enter ISO file path: " IMG
+fi
+if [ ! -f "$IMG" ]; then
+    echo "Usage: $0 <path_to_iso_file>"
+    exit 1
+fi
+
 if ! dpkg -l $pkg > /dev/null 2>&1; then
     sudo apt install $pkg -y
 fi
@@ -15,14 +24,13 @@ fi
 
 # absolute path of the ISO file
 IMG=$(realpath "$IMG")
-
-#  ISO file name without the suffix
-basename=$(basename "$IMG" "$suffix")
+basename=$(basename "$IMG")
 dirname=$(dirname "$IMG")
 
 # output folder, zip file, and checksum file name
-folder="$dirname/$basename"
-zipfile="$dirname/$basename.zip"
+name_without_ext="${basename%.*}"
+folder="$dirname/$name_without_ext"
+zipfile="$dirname/$name_without_ext.zip"
 checksumfile="$zipfile.sha256sum"
 
 rm -rf "$folder"
@@ -38,5 +46,5 @@ cd "$dirname" || exit
 sha256sum "$zipfile" > "$checksumfile"
 
 echo ""
-echo "The checksum of $zipfile is:"
+echo "ZIP completed.The checksum of $zipfile is:"
 cat "$checksumfile"
